Polygons

In RationalPolygons.jl, we represent a polygon $P \subseteq \mathbb{R}^2$ by two pieces of data: An integral matrix $V \in \mathbb{Z}^{2\times N}$, called the vertex matrix and an integer $k \in \mathbb{Z}$, called the rationality. The associated polygon has as vertices the columns of $V$ divided by $k$. To represent $V$, we use static matrices, which are faster than Julia's internal matrices for many common operations. However, this implies that the type RationalPolygon depends on the number of vertices $N$ as a type parameter, which affects Julia's dispatch mechanism at runtime. As long as the number of distinct values of $N$ occurring during a computation remains relatively small, this should not cause performance issues.

There are two ways in which this encoding of rational polygons is not unique: First, scaling $V$ and $k$ by the same factor does not change the polygon, e.g. $(V,k)$ describes the same polygon as $(2V,2k)$. Even though they are mathematically the same polygon, RationalPolygon.jl views them as different objects: once as a $k$-rational polygon and once as a $2k$-rational polygon. The second way in which this encoding is not unique is that we can change the order of the columns. While we require them to be sorted counterclockwise, we may use any vertex as the first column.

RationalPolygons.RationalPolygon — Type

RationalPolygon{T<:Integer,N,M}

The type of rational polygons in two-dimensional space. T is the type of integers to be used. $N$ is the number of vertices of the polygon and $M$ equals $2 \cdot N$. It has the following fields:

rationality :: T: The rationality of the polygon, e.g. $1$ for lattice polygons, $2$ for half-integral polygons etc.

vertex_matrix :: SMatrix{2,N,T,M}: An integral $2 \times N$ matrix. The vertices of the polygon are understood to be the columns of this matrix divided by rationality.

number_of_vertices :: Int: The number of vertices of the polygon. This is redundant information, since the number of vertices is already available as the type parameter $N$. However, getting the number of vertices of a polygon through the type parameter means lots of work for Julia's dispatch algorithm. Therefore, we found it to improve performance to put it as a variable into the struct as well.

is_unimodular_normal_form :: Bool: A flag to remember that the polygon is already in unimodular normal form. This is used internally to avoid redundant computations of the normal form.

is_affine_normal_form :: Bool: A flag variable to remember that the polygon is already in affine normal form. This is used internally to avoid redundant computations of the normal form.

The rest of this section is organized as follows. In Subsection $\ref{doc:Constructors}$, we describe different methods of constructing a rational polygon. Subsection $\ref{doc:Basic-Properties}$ is about basic properties. In Subsection $\ref{doc:Ehrhart-Theory}$, we discuss various ways of counting lattice points as well as Ehrhart theory. Subsection $\ref{doc:LDP-polygons-and-toric-surfaces}$ is about properties of LDP polygons that correspond to meaningful invariants of their associated toric del Pezzo surfaces. In Subsection $\ref{doc:Normal-forms-and-automorphism-groups}$, we discuss unimodular and affine unimodular normal forms. Finally, Subsection $\ref{doc:Lattice-width}$ is about lattice width and related concepts.

Constructors

To construct a rational polygon, one can either use a type constructor or one of the functions convex_hull and intersect_halfplanes. We describe the type constructors first.

RationalPolygons.RationalPolygon — Method

RationalPolygon(vertex_matrix :: SMatrix{2,N,T,M}, rationality :: T) where {N, M, T <: Integer}
RationalPolygon(scaled_vertices :: Vector{LatticePoint{T}}, rationality :: T)
RationalPolygon(vertices :: Vector{RationalPoint{T}})
RationalPolygon(vertices :: Vector{RationalPoint{T}}, rationality :: T)

Construct a rational polygon from a given set of vertices and possibly prescribed rationality. If vertex_matrix or scaled_vertices is provided, these are understood to be the integral vertices of the scaled polygon rationality(P) * P.

These constructors don't perform any consistency checks on the input. In particular, the user must be sure that the given points are truly vertices of the polygon and that they are ordered counterclockwise. If this is not known ahead of construction, convex_hull should be used instead.

All constructors accept the optional keyword arguments is_unimodular_normal_form and is_affine_normal_form, which are set to false by default. If they are set to true, this means the user is certain that the given polygon is already in the respective normal form. This information will be used to prevent additional computations of the normal form and thus speed up equivalence checking.

Example:

The standard lattice triangle.

julia> P = RationalPolygon(LatticePoint{Int}[(0,0),(1,0),(0,1)], 1)
Rational polygon of rationality 1 with 3 vertices.   

Example:

A half-integral polygon.

julia> P = RationalPolygon(RationalPoint{Int}[(1,0),(0,1//2),(-1,0),(0,-1//2)])
Rational polygon of rationality 2 with 4 vertices.

RationalPolygons.convex_hull — Function

convex_hull(points :: Vector{LatticePoint{T}}, k :: T = one(T)) where {T <: Integer}

Return the $k$-rational polygon given by the convex hull of $p / k$, where p ∈ points.

convex_hull(points :: Vector{RationalPoint{T}}) where {T <: Integer}

Return the convex hull of a given set of rational points. The rationality will be inferred from the input.

Example:

julia> convex_hull(RationalPoint{Int}[(1,0),(0,1//2),(0,-1//3)])
Rational polygon of rationality 6 with 3 vertices.

convex_hull(points :: Vector{RationalPoint{T}}, k :: T) where {T <: Integer}

Return the convex hull of a given set of rational points, viewed as a $k$-rational polygon.

Example:

julia> convex_hull(RationalPoint{Int}[(1,0),(0,1//2),(0,-1//3)], 12)
Rational polygon of rationality 12 with 3 vertices.

RationalPolygons.intersect_halfplanes — Function

intersect_halfplanes(halfplanes :: Vector{AffineHalfplane{T}}; rationality :: Union{Missing,T} = missing) where {T <: Integer}

Return the polygon described by the intersection of the given affine halfplanes. Throws an error if the intersection is unbounded. If rationality is not provided, it will be inferred from the input.

RationalPolygons.empty_polygon — Function

empty_polygon(rationality :: T) where {T <: Integer}

Return the empty polygon of given rationality.

empty_polygon(::Type{T}) where {T <: Integer}

Return the empty polygon of integer type T. The rationality is understood to be one, i.e. it is a lattice polygon.

Basic Properties

We provide basic properties and checks for rational polygons. Note that indices corresponding to vertices are always considered cyclic, i.e. the $N+1$-th vertex of a polygon with $N$ vertices equals its first vertex.

RationalPolygons.number_of_vertices — Function

number_of_vertices(P :: RationalPolygon)

Return the number of vertices of a polygon.

RationalPolygons.rationality — Method

rationality(P :: RationalPolygon)

Return the rationality of $P$. Note that this does not need to be the denominator of $P$, which is the smallest positive integer $k$ such that $kP$ is a lattice polygon: The standard lattice triangle may also be viewed as a half-integral polygon, in which case the rationality would be two, but the denominator is one.

Base.denominator — Method

Base.denominator(P :: RationalPolygon{T,N}) where {N,T <: Integer}

Return the smallest positive integer $k$ such that $kP$ is a lattice polygon.

RationalPolygons.vertex_matrix — Function

vertex_matrix(P :: RationalPolygon)

Return the $2 \times N$ integral matrix containing the vertices of rationality(P) * P as its columns.

Example:

julia> P = convex_hull(RationalPoint{Int}[(1,0),(0,1//2),(0,-1//3)])
Rational polygon of rationality 6 with 3 vertices.

julia> vertex_matrix(P)
2×3 SMatrix{2, 3, Int64, 6} with indices SOneTo(2)×SOneTo(3):
  0  6  0
 -2  0  3

RationalPolygons.scaled_vertex — Function

scaled_vertex(P :: RationalPolygon, i :: Int)

Return the $i$-th integral vertex of the polygon rationality(P) * P. The index $i$ is regarded as a cyclic index, e.g. the $N+1$-th vertex is equal to the first vertex.

RationalPolygons.vertex — Function

vertex(P :: RationalPolygon, i :: Int)

Return the $i$-th vertex of the polygon $P$. The index $i$ is regarded as a cyclic index, e.g. the $N+1$-th vertex is equal to the first vertex.

RationalPolygons.vertices — Function

vertices(P :: RationalPolygon)

Return the vertices of $P$.

RationalPolygons.affine_halfplane — Method

affine_halfplane(P :: RationalPolygon, i :: Int)

Return the $i$-th describing halfplane of $P$.

RationalPolygons.affine_halfplanes — Method

affine_halfplanes(P :: RationalPolygon)

Return the describing halfplanes of $P$.

Base.in — Method

Base.in(x :: Point{T}, P :: RationalPolygon{T}) where {T <: Integer}

Check whether a point $x$ is contained in the polygon $P$.

Example:

julia> P = convex_hull(LatticePoint{Int}[(1,0),(0,1),(-1,-1)])
Rational polygon of rationality 1 with 3 vertices.

julia> Point(0,0) ∈ P
true

julia> Point(1//2, 1//2) ∈ P
true

RationalPolygons.contains_in_interior — Method

contains_in_interior(x :: Point{T}, P :: RationalPolygon{T}) where {T <: Integer}

Check whether a point $x$ is contained in the interior of $P$.

Example:

julia> P = convex_hull(LatticePoint{Int}[(1,0),(0,1),(-1,-1)])
Rational polygon of rationality 1 with 3 vertices.

julia> contains_in_interior(Point(0,0), P)
true

julia> contains_in_interior(Point(1//2,1//2), P)
false

RationalPolygons.dim — Function

dim(P :: RationalPolygon)

Return the dimension of $P$. For empty polygons, this returns $-1$. Otherwise, it returns $0$, $1$ or $2$.

RationalPolygons.normalized_area — Function

normalized_area(P :: RationalPolygon)

Return the normalized area of a $k$-rational polygon, i.e. $2k^2$ times its euclidean area. The result is always an integer, counting the number of standard $k$-rational triangles contained in $P$.

RationalPolygons.euclidean_area — Function

euclidean_area(P :: RationalPolygon)

Return the euclidean area of a rational polygon.

RationalPolygons.is_maximal — Function

is_maximal(P :: RationalPolygon)

Check whether a rational polygon is maximal among all polygons sharing the same rationality and number of interior lattice points.

Example:

julia> P = convex_hull(LatticePoint{Int}[(0,0),(3,0),(0,3)])
Rational polygon of rationality 1 with 3 vertices.

julia> is_maximal(P)
true

julia> Q = convex_hull(LatticePoint{Int}[(0,0),(2,0),(2,1),(0,3)])
Rational polygon of rationality 1 with 4 vertices.

julia> is_maximal(Q)
false

RationalPolygons.dual — Function

dual(P :: RationalPolygon{T}) where {T <: Integer}

Return the dual of a polygon $P$. Throws an error if $P$ does not contain the origin in its interior.

Example:

julia> P = convex_hull(LatticePoint{Int}[(-1,-1),(1,0),(0,1)])
Rational polygon of rationality 1 with 3 vertices.

julia> vertex_matrix(dual(P))
2×3 SMatrix{2, 3, Int64, 6} with indices SOneTo(2)×SOneTo(3):
 -2   1  1
  1  -2  1

Ehrhart Theory

Recall that the number of lattice points in integral multiples of a $k$-rational polygon $P$ is a quasipolynomial, called its Ehrhart quasipolynomial:

\[\mathrm{ehr_P}(t) = |tP \cap \mathbb{Z}^2| = At^2 + a(t)t+b(t), \qquad t \in \mathbb{Z}.\]

Here, $A$ is the euclidean area of $P$ and $a, b\colon \mathbb{Z} \to \mathbb{Q}$ are $k$-periodic functions. These can be computed by

\[a(t) = -(2t+k)\cdot A + \frac{\mathrm{ehr}_P(t+k)-\mathrm{ehr}_P(t)}{k},\]

\[b(t) = (t^2+tk)\cdot A + \frac{(t+k)\mathrm{ehr}_P(t)-t\mathrm{ehr}_P(t+k)}{k}.\]

Setting $\tilde{A} := 2k^2A,\ \tilde{a} := 2k^2 a$ and $\tilde{b} := 2k^2b$, we get integer-valued functions $\tilde a$ and $\tilde b$, which we call the normalized Ehrhart coefficients. In RationalPolygons.jl, we encode the Ehrhart quasipolynomial by the $3\times k$-integral matrix of its normalized Ehrhart coefficients:

\[\begin{bmatrix} \tilde{A} & \tilde{a}(1) & \tilde{b}(1) \\ \tilde{A} & \tilde{a}(2) & \tilde{b}(2) \\ \vdots & \vdots & \vdots \\ \tilde{A} & \tilde{a}(k) & \tilde{b}(k) \end{bmatrix}\]

If $P$ is integral, we have $k=1$ and its Ehrhart quasipolynomial is a regular polynomial of degree two. In general, the periods of $a$ and $b$ are divisors of $k$.

RationalPolygons.jl comes with many methods for counting the (interior, boundary) lattice points of a rational polygon as well as computing its Ehrhart quasipolynomial and its periods.

Counting lattice points

RationalPolygons.generic_lattice_points — Function

generic_lattice_points(P :: RationalPolygon{T,N}, k :: T)

Return the lattice points in the $k$-fold of the rational polygon $P$. If the keyword argument interior = true is passed, only the lattice points in the relative interior are computed. If only_count = true is passed, only the total number of lattice points is returned.

RationalPolygons.boundary_k_rational_points — Function

boundary_k_rational_points(P :: RationalPolygon{T,N}, k :: T) where {N,T <: Integer}

Return all $k$-rational points on the boundary of $P$.

RationalPolygons.number_of_boundary_k_rational_points — Function

number_of_boundary_k_rational_points(P :: RationalPolygon{T}, k :: T) where {T <: Integer}

Return the number of $k$-rational points on the boundary of $P$.

RationalPolygons.boundary_lattice_points — Function

boundary_lattice_points(P :: RationalPolygon{T}) where {T <: Integer}

Return all lattice points on the boundary of $P$.

RationalPolygons.number_of_boundary_lattice_points — Function

number_of_boundary_lattice_points(P :: RationalPolygon)

Return the number of lattice points on the boundary of $P$.

RationalPolygons.interior_k_rational_points — Function

interior_k_rational_points(P :: RationalPolygon{T}, k :: T) where {T <: Integer}

Return all $k$-rational points in the interior of $P$.

RationalPolygons.number_of_interior_k_rational_points — Function

number_of_interior_k_rational_points(k :: T, P :: RationalPolygon{T}) where {T <: Integer}

Return the number of $k$-rational points in the interior of $P$.

RationalPolygons.interior_lattice_points — Function

interior_lattice_points(P :: RationalPolygon{T}) where {T <: Integer}

Return all lattice points in the interior of $P$.

RationalPolygons.number_of_interior_lattice_points — Function

number_of_interior_lattice_points(P :: RationalPolygon{T}) where {T <: Integer}

Return the number of lattice points in the interior of $P$.

RationalPolygons.k_rational_points — Function

k_rational_points(k :: T, P :: RationalPolygon{T}) where {T <: Integer}

Return all $k$-rational points in $P$.

RationalPolygons.number_of_k_rational_points — Function

number_of_k_rational_points(P :: RationalPolygon{T}, k :: T) where {T <: Integer}

Return the number of $k$-rational points in $P$.

RationalPolygons.lattice_points — Function

lattice_points(P :: RationalPolygon{T}) where {T <: Integer}

Return all lattice points in $P$.

RationalPolygons.number_of_lattice_points — Function

number_of_lattice_points(P :: RationalPolygon{T}) where {T <: Integer}

Return the number of lattice points in $P$.

RationalPolygons.k_rational_hull — Function

k_rational_hull(P :: RationalPolygon{T}, k :: T) where {T <: Integer}

Return the convex hull of all $k$-rational points contained in $P$. If primitive = true is passed, only the primitive $k$-rational points are taken.

RationalPolygons.interior_k_rational_hull — Function

interior_k_rational_hull(P :: RationalPolygon{T}, k :: T) where {T <: Integer}

Return the convex hull of all interior $k$-rational points of $P$. If primitive = true is passed, only the primitive $k$-rational points are taken.

RationalPolygons.integer_hull — Function

integer_hull(P :: RationalPolygon{T}) where {T <: Integer}

Return the convex hull of all lattice points in the interior of $P$. If primitive = true is passed, only the primitive interior lattice points are taken.

RationalPolygons.interior_integer_hull — Function

interior_integer_hull(P :: RationalPolygon{T}) where {T <: Integer}

Return the convex hull of all interior lattice points of $P$. If primitive = true is passed, only the primitive lattice points are taken.

Ehrhart quasipolynomial

RationalPolygons.is_periodic — Function

is_periodic(v :: Vector, k :: Int)

Check if a vector $v$ is $k$-periodic, i.e. v[i] == v[mod(i+k,1:n)] for all i, where n = length(v).

RationalPolygons.period — Function

period(v :: Vector)

Return the smallest positive integer $k$ such that $v$ is $k$-periodic.

RationalPolygons.ehrhart_quasipolynomial_with_periods — Function

ehrhart_quasipolynomial_with_periods(P :: RationalPolygon{T}) where {T <: Integer}

Return a $k \times 3$-matrix of normalized coefficients of the Ehrhart quasipolynomial of a $k$-rational polygon $P$, together with a vector of its three periods.

RationalPolygons.ehrhart_quasipolynomial — Function

ehrhart_quasipolynomial(P :: RationalPolygon)

Return a $k \times 3$-matrix of normalized coefficients of the Ehrhart quasipolynomial of a $k$-rational polygon $P$.

RationalPolygons.ehrhart_quasipolynomial_periods — Function

ehrhart_quasipolynomial_periods(P :: RationalPolygon)

Return the periods of the Ehrhart quasipolynomial of a $k$-rational polygon $P$.

RationalPolygons.ehrhart_quasipolynomial_period — Function

ehrhart_quasipolynomial_period(P :: RationalPolygon)

Return the period of the Ehrhart quasipolynomial of a $k$-rational polygon $P$.

RationalPolygons.is_quasiintegral — Function

is_quasiintegral(P :: RationalPolygon)

Check whether a rational polygon is quasiintegral, i.e. has an Ehrhart polynomial.

LDP polygons and toric surfaces

An LDP polygon is a lattice polygon with primitive vertices containing the origin in its interior. LDP polygons correspond to toric log del Pezzo surfaces. In this package, we call more generally a $k$-rational polygon LDP, if it contains the origin in its interior and its $k$-fold multiple has primitive vertices. With this definition, the $k$-rational LDP polygons with exactly one interior lattice point correspond to the toric del Pezzo surfaces with at most $\frac{1}{k}$-log canonical singularities. Here, we list some properties of LDP polygons that correspond to meaningful invariants of the associated toric del Pezzo surface. For more background on LDP polygons, see Section $\ref{sec:ldp_polygons_toric_log_del_pezzo_surfaces}$.

RationalPolygons.contains_origin_in_interior — Function

contains_origin_in_interior(P :: RationalPolygon)

Check whether $P$ contains the origin in its interior.

RationalPolygons.is_primitive — Method

is_primitive(P :: RationalPolygon)

Check whether all vertices of the scaled lattice polygon rationality(P) * P are primitive.

RationalPolygons.is_ldp — Function

is_ldp(P :: RationalPolygon)

Check whether $P$ is LDP, i.e. is primitive and contains the origin in its interior.

RationalPolygons.multiplicity — Method

multiplicity(P :: RationalPolygon{T}, i :: Int) where {T <: Integer}

The index of the sublattice spanned by the $i$-th and $i+1$-th scaled vertex of $P$ (i.e. the determinant of those two vertices). For LDP polygons, this equals the order of the local class group associated with the toric fixed point associated to the $i$-th and $i+1$-th ray. See Definition $\ref{def:ldp_polygon_multiplicity}$ and Proposition $\ref{prp:ldp_polygon_local_multiplicy_grading_matrix}$.

RationalPolygons.multiplicity — Method

multiplicity(P :: RationalPolygon{T,N}) where {N,T <: Integer}

The order of the sublattice spanned by the scaled vertices of $P$. For LDP polygons, this equals the order of the torsion part of the divisor class group of the associated toric surface. See Definition $\ref{def:ldp_polygon_multiplicity}$ and Proposition $\ref{prp:class_group_splitting}$.

RationalPolygons.grading_matrix_free_part — Function

grading_matrix_free_part(P :: RationalPolygon)

Return the free part of the grading matrix associated to $P$.

Example:

julia> P = convex_hull(LatticePoint{Int}[(1,0), (2,5), (-4,-5), (-1,-5)])
Rational polygon of rationality 1 with 4 vertices.

julia> grading_matrix_free_part(P)
2×4 SMatrix{2, 4, Int64, 8} with indices SOneTo(2)×SOneTo(4):
 1  -1  3  0
 1   0  2  1

RationalPolygons.grading_matrix_torsion_part — Function

grading_matrix_torsion_part(P :: RationalPolygon)

Return the torsion part of the grading matrix associated to $P$. If $P$ has $N$ vertices and has multiplicity $\mu$, the result is a static vector of length $N$ whose entries are all between $0$ and $\mu-1$.

Example:

julia> P = convex_hull(LatticePoint{Int}[(1,0), (2,5), (-4,-5), (-1,-5)])
Rational polygon of rationality 1 with 4 vertices.

julia> multiplicity(P)
5

julia> grading_matrix_torsion_part(P)
4-element SVector{4, Int64} with indices SOneTo(4):
 4
 0
 1
 0

RationalPolygons.grading_matrix — Function

grading_matrix(P :: RationalPolygon{T,N}) where {N, T <: Integer}

Return a tuple $(Q_0, Q_1)$ where $Q_0$ is the free part and $Q_1$ the torsion part of the grading matrix associated to $P$.

RationalPolygons.is_smooth — Function

is_smooth(P :: RationalPolygon, i :: Int)

Check whether the cone spanned by the $i$-th and $i+1$-th vertex of $P$ is regular, i.e. generates the entire lattice. For LDP polygons, this means that the toric fixed point associated to the $i$-th and $i+1$-th ray is smooth.

is_smooth(P :: RationalPolygon)

Check whether all cones of the face fan of $P$ are regular. For LDP polygons, this means that the associated toric surface is smooth.

RationalPolygons.picard_index — Function

picard_index(P :: RationalPolygon{T,N}) where {N,T <: Integer}

The product of all local multiplicities of $P$ divided by the global multiplicity. For LDP polygons, this equals the index of the Picard group inside the divisor class group of the associated toric surface, see Theorem $\ref{thm:picard_index_formula_ldp_polygons}$

RationalPolygons.gorenstein_index — Function

gorenstein_index(P :: RationalPolygon{T}, i :: Int)

The multiplicity of $P$ divided by gcd(w[2] - v[2], v[1] - w[1]), where $v$ and $w$ are the $i$-th and $i+1$-th scaled vertices of $P$ respectively. For LDP polygons, this equals the local Gorenstein at the toric fixed point associated to the $i$-th and $i+1$-th ray of $P$, see Proposition $\ref{prp:gorenstein_index_formula}$

gorenstein_index(P :: RationalPolygon{T}) where {T <: Integer}

The least common multiple of the local Gorenstein indices of $P$. For LDP polygons, this equals the Gorenstein index of the associated toric surface.

RationalPolygons.gorenstein_coefficients — Function

gorenstein_coefficients(P :: RationalPolygon{T,N})

Return the Gorenstein coefficients of an LDP polygon with $N$ vertices. This is an integral matrix $A = (a_{ij}) \in \ZZ^{N \times (N-2)}$ such that $\iota w = \sum_{j=1}^{N-2} a_{ij} w_{j+i-1}$, where $\iota$ is the Gorenstein index, $w_i$ are the columns of the free part of the grading matrix, and $w = w_1 + \dots + w_N$ is the class of the anticanonical divisor. See Construction $\ref{cns:picard_group_of_free_part_grading_matrix}$.

Example:

julia> P = convex_hull(LatticePoint{Int}[(1,0), (2,5), (-4,-5), (-1,-5)])
Rational polygon of rationality 1 with 4 vertices.

julia> gorenstein_coefficients(P)
4×2 SMatrix{4, 2, Int64, 8} with indices SOneTo(4)×SOneTo(2):
 20   5
 15  10
  5  10
  5  15

RationalPolygons.gorenstein_matrix — Function

gorenstein_matrix(P :: RationalPolygon{T,N}) where {T <: Integer, N}

Return the Gorenstein matrix of an LDP polygon. See Definition $\ref{def:gorenstein_matrix}$.

Example:

julia> P = convex_hull(LatticePoint{Int}[(1,0), (2,5), (-4,-5), (-1,-5)])
Rational polygon of rationality 1 with 4 vertices.

julia> gorenstein_matrix(P)
4×4 SMatrix{4, 4, Int64, 16} with indices SOneTo(4)×SOneTo(4):
 -15    0   5   5
   5  -10  -5   5
   5    5   0  -5
 -10    5   5   0

RationalPolygons.discrepancies — Function

discrepancies(v1 :: LatticePoint{T}, v2 :: LatticePoint{T}) where {T <: Integer}

The discrepancies of the two-dimensional cone spanned by two given integral primitive vectors. See Definition $\ref{def:local_log_canonicity}$ and Proposition $\ref{prp:local_log_canonicity_formula}$.

RationalPolygons.log_canonicity — Function

log_canonicity(P :: RationalPolygon)

Given a $k$-rational polygon $P$, return the maximal rational number $0 < \varepsilon \leq 1$ such that $\varepsilon*P$ contains only one $k$-rational point in its interior (the origin). For an LDP polygon, this equals the maximal rational number $0 < \varepsilon \leq 1$ such that the associated toric surface is $\varepsilon$-log canonical. See Definition $\ref{def:ldp_polygon_log_canonical}$.

RationalPolygons.toric_prime_divisor_self_intersection — Function

toric_prime_divisor_self_intersection(P :: RationalPolygon, i :: Int)

Writing $u$, $v$ and $w$ for the $i-1$-th, $i$-th and $i+1$-th scaled vertex of $P$ respectively, return $\frac{det(w,u)}{det(u,v) det(v,w)}$. For LDP polygons, this equals the self intersection number of the $i$-th toric prime divisor. See Definition $\ref{def:ldp_polygon_intersection_numbers}$.

RationalPolygons.toric_prime_divisor_adjacent_intersection — Function

toric_prime_divisor_adjacent_intersection(P :: RationalPolygon, i :: Int)

Writing $v$ and $w$ for the $i$-th and $i+1$-th scaled vertex of $P$, return $\frac{1}{\mathrm{det}(v,w)}$. For LDP polygons, this equals the intersection number between the $i$-th and $i+1$-th toric prime divisors. See Definition $\ref{def:ldp_polygon_intersection_numbers}$.

RationalPolygons.degree — Method

degree(P :: RationalPolygon)

For LDP polygons, return the self intersection number of an anticanonical divisor of the associated toric surface. See Definition $\ref{def:ldp_polygon_intersection_numbers}$.

Example:

The projective plane has degree 9.

julia> P = convex_hull(LatticePoint{Int}[(1,0),(0,1),(-1,-1)])
Rational polygon of rationality 1 with 3 vertices.

julia> degree(P)
9//1

Normal forms and automorphism groups

Two $k$-rational polygons are called (affine) unimodular equivalent if they can be transformed into each other by an (affine) unimodular transformation. The purpose of a normal form is to provide a unique representative for every equivalence class, i.e. two polygons should be (affine) unimodular equivalent to each other if and only if their (affine) unimodular normal forms coincide.

For details about the normal form used in RationalPolygons.jl, see Section $\ref{subsec:normal_forms}$.

RationalPolygons.unimodular_normal_form — Function

unimodular_normal_form(P :: RationalPolygon{T,N}) where {N,T <: Integer}

Return a unimodular normal form of a rational polygon. Two rational polygons have the same unimodular normal form if and only if the can be transformed into each other by applying a unimodular transformation.

RationalPolygons.are_unimodular_equivalent — Function

are_unimodular_equivalent(P :: RationalPolygon, Q :: RationalPolygon)

Check whether two rational polygons are equivalent by a unimodular transformation.

RationalPolygons.affine_normal_form — Function

affine_normal_form(P :: RationalPolygon{T,N}) where {N,T <: Integer}

Return an affine normal form of a rational polygon. Two rational polygons have the same affine normal form if and only if the can be transformed into each other by applying an affine unimodular transformation.

RationalPolygons.are_affine_equivalent — Function

are_affine_equivalent(P :: RationalPolygon, Q :: RationalPolygon)

Check whether two rational polygons are equivalent by an affine unimodular transformation.

RationalPolygons.PolygonAutomorphismGroup — Type

PolygonAutomorphismGroup

A struct holding information about the automorphism group of a rational polygon. It has two fields is_cyclic :: Bool and n :: Int.

RationalPolygons.CyclicGroup — Function

CyclicGroup(n :: Int)

Return the cyclic group of order $n$, as a PolygonAutomorphismGroup.

RationalPolygons.DihedralGroup — Function

DihedralGroup(n :: Int)

Return the dihedral group of order $2n$, as a PolygonAutomorphismGroup.

RationalPolygons.is_cyclic — Function

is_cyclic(G :: PolygonAutomorphismGroup)

Check whether $G$ is a cyclic group.

RationalPolygons.order — Function

order(G :: PolygonAutomorphismGroup)

Return the order of the group $G$.

RationalPolygons.unimodular_automorphism_group — Function

unimodular_automorphism_group(P :: RationalPolygon)

Return the automorphism group of $P$ with respect to unimodular transformations.

Example:

julia> P = convex_hull(LatticePoint{Int}[(1,0),(0,1),(-1,1),(-1,0),(0,-1),(1,-1)])
Rational polygon of rationality 1 with 6 vertices.

julia> unimodular_automorphism_group(P)
D6

RationalPolygons.affine_automorphism_group — Function

affine_automorphism_group(P :: RationalPolygon{T,N}) where {N,T <: Integer}

Return the automorphism group of $P$ with respect to affine unimodular transformations.

Lattice width

We provide functions to compute the lattice width as well as all direction vectors in which the lattice width is attained. Furthermore, we implement the concept of lattice width data following Bohnert [3], which captures information about the slicing lengths of a polygon with respect to given direction vectors.

RationalPolygons.width — Function

width(P :: RationalPolygon{T}, w :: Point{T}) where {T <: Integer}

Return the lattice width of $P$ in direction $w$.

Example:

julia> P = convex_hull(LatticePoint{Int}[(1,1),(1,-2),(-4,2),(-2,2)],2)
Rational polygon of rationality 2 with 4 vertices.

julia> width(P, Point(0,1))
2//1

julia> width(P, Point(1,0))
5//2

width(P :: RationalPolygon)

Return the lattice width of $P$.

Example:

julia> P = convex_hull(LatticePoint{Int}[(1,1),(1,-2),(-4,2),(-2,2)],2)
Rational polygon of rationality 2 with 4 vertices.

julia> width(P)
2//1

RationalPolygons.all_direction_vectors_with_width_less_than — Function

all_direction_vectors_with_width_less_than(P :: RationalPolygon{T}, c :: Rational{T}) where {T <: Integer}

Return all direction vectors in which the width of $P$ is less than or equal to a given constant.

Example:

julia> P = convex_hull(LatticePoint{Int}[(1,1),(1,-2),(-4,2),(-2,2)],2)
Rational polygon of rationality 2 with 4 vertices.

julia> all_direction_vectors_with_width_less_than(P, 3//1)
4-element Vector{SVector{2, Int64}}:
 [1, 0]
 [0, 1]
 [1, 1]
 [1, 2]

RationalPolygons.width_direction_vectors — Function

width_direction_vectors(P :: RationalPolygon)

Return the lattice width direction vectors of $P$, i.e. those directions that realize the lattice width of $P$.

Example:

julia> P = convex_hull(LatticePoint{Int}[(1,1),(1,-2),(-4,2),(-2,2)],2)
Rational polygon of rationality 2 with 4 vertices.

julia> width_direction_vectors(P)
2-element Vector{SVector{2, Int64}}:
 [0, 1]
 [1, 1]

RationalPolygons.adjust_to_width_direction — Function

adjust_to_width_direction(P :: RationalPolygon{T}, w :: Point{T}) where {T <: Integer}

Apply an affine unimodular transformation to $P$ that transforms the given width direction vector to $(1,0)$, see Lemma 2.10 of [3].

RationalPolygons.number_of_interior_integral_lines — Function

number_of_interior_integral_lines(P :: RationalPolygon{T}, w :: Point{T}) where {T <: Integer}

Return the number of interior integral lines of $P$ with respect to a given direction vector $w$.

Example:

julia> P = convex_hull(RationalPoint{Int}[(1//3,-1),(4//3,2),(2//3,2),(-4//3,-1)])
Rational polygon of rationality 3 with 4 vertices.

julia> number_of_interior_integral_lines(P, LatticePoint(1,0))
3

julia> number_of_interior_integral_lines(P, LatticePoint(0,1))
2

RationalPolygons.minimal_number_of_interior_integral_lines — Function

minimal_number_of_interior_integral_lines(P :: RationalPolygon{T}) where {T <: Integer}

Return the minimal number of interior integral lines of $P$.

Example:

julia> P = convex_hull(RationalPoint{Int}[(1//3,-1),(4//3,2),(2//3,2),(-4//3,-1)])
Rational polygon of rationality 3 with 4 vertices.

julia> minimal_number_of_interior_integral_lines(P)
2

RationalPolygons.is_realizable_in_interval — Function

is_realizable_in_interval(P :: RationalPolygon{T}, h :: T) where {T <: Integer}

Check whether $P$ is realizable in $\mathbb{R} \times [0,h]$. This is true if and only if the minimal number of interior integral lines is less than or equal to $h-1$.

Example:

julia> P = convex_hull(RationalPoint{Int}[(1//3,-1),(4//3,2),(2//3,2),(-4//3,-1)])
Rational polygon of rationality 3 with 4 vertices.

julia> is_realizable_in_interval(P,2)
false

julia> is_realizable_in_interval(P,3)
true

RationalPolygons.LatticeWidthData — Type

LatticeWidthData{T <: Integer}

A struct capturing information about the lattice width of a rational polygon with respect to a lattice width direction vector, see Definition 2.11 of [3]. It has two fields:

interval_of_nonzero_vertical_slice_length :: Tuple{Rational{T},Rational{T}},

interval_of_longest_vertical_slice_length :: Tuple{Rational{T},Rational{T}}.

RationalPolygons.lattice_width_data — Function

lattice_width_data(P :: RationalPolygon{T}, w :: Point{T}) where {T <: Integer}

Compute the lattice width data of a rational polygon with respect to a given width direction vector, see Definition 2.11 of [3]. This function returns a value of type LatticeWidthData.

Example:

julia> P = convex_hull(LatticePoint{Int}[(1,1),(1,-2),(-4,2),(-2,2)],2)
Rational polygon of rationality 2 with 4 vertices.

julia> ws = width_direction_vectors(P)
2-element Vector{SVector{2, Int64}}:
 [0, 1]
 [1, 1]

julia> lattice_width_data(P,ws[1])
LatticeWidthData{Int64}((0//1, 2//1), (3//2, 3//2))

julia> lattice_width_data(P,ws[2])
LatticeWidthData{Int64}((0//1, 2//1), (1//2, 1//2))

RationalPolygons.number_of_interior_integral_vertical_lines — Function

number_of_interior_integral_vertical_lines(P :: RationalPolygon, w :: Point{T}) where {T <: Integer}

Return the number of interior integral vertical lines of $P$ with respect to a given lattice width direction vector, see Definition 2.11 of [3].

RationalPolygons.position_of_longest_vertical_slice_length — Function

position_of_longest_vertical_slice_length(P :: RationalPolygon, w :: Point{T}) where {T <: Integer}

Return the position of the longest vertical slicing length of $P$ with respect to a given lattice width direction vector, see Definition 2.11 of [3].

RationalPolygons.lattice_width_datas — Function

lattice_width_datas(P :: RationalPolygon{T}) where {T <: Integer}

Return the lattice width data for all lattice width direction vectors of $P$.

RationalPolygons.numbers_of_interior_integral_vertical_lines — Function

numbers_of_interior_integral_vertical_lines(P :: RationalPolygon{T}) where {T <: Integer}

Return the number of interior integral vertical lines for all width direction vectors of $P$.

RationalPolygons.positions_of_longest_vertical_slice_length — Function

positions_of_longest_vertical_slice_length(P :: RationalPolygon{T}) where {T <: Integer}

Return the positions of the longest vertical slicing lengths for all width direction vectors of $P$.

IO

RationalPolygons.jl provides two ways to read and write polygons from files. The first is text-based. Polygons can be read from files containing one polygon per line like this:

[[2, 0], [1, 3], [-1, 0], [-3, -4]]
[[1, 0], [2, 6], [-4, -9]]
[[1, 0], [3, 5], [0, 1], [-5, -8]]
....

This text-based format has the advantage of being easy to understand and use. However, storing polygons as ASCII strings is not very space-efficient, as they contain many redundant control characters. Hence we provide another way to store polygons in binary and compressed form, which uses the HDF5 file format. This is more suitable for large datasets. For an example session, see write_polygon_dataset.

RationalPolygons.parse_rational_polygons — Function

parse_rational_polygons(k :: T, files :: AbstractVector{String}) where {T <: Integer}
parse_rational_polygons(k :: T, file :: String) where {T <: Integer}

Parse a list of files containing the vertices of a $k$-rational polygon. The files must have one polygon per line containing its vertices, as in the following example:

[[2, 0], [1, 3], [-1, 0], [-3, -4]]
[[1, 0], [2, 6], [-4, -9]]
[[1, 0], [3, 5], [0, 1], [-5, -8]]
....

RationalPolygons.write_rational_polygons — Function

write_rational_polygons(Ps :: Vector{<:RationalPolygon{T}}, filepath :: String) where {T <: Integer}

Write a list of polygons to a text file. Each polygon will be written as one line containing its vertices, as in the following example:

[[2, 0], [1, 3], [-1, 0], [-3, -4]]
[[1, 0], [2, 6], [-4, -9]]
[[1, 0], [3, 5], [0, 1], [-5, -8]]
....

RationalPolygons.create_polygon_dataset — Function

create_polygon_dataset(f :: Union{HDF5.File, HDF5.Group}, path :: String, N :: Int)

Create an HDF5 dataset named path for storing rational polygons with $N$ vertices. The dataset will have an HDF5 compound datatype with $2N$ integers, which are the entries of the vertex matrices stored in a column major layout. This function takes three keyword arguments:

T :: Type{<:Integer}: The integer type to be used, e.g. Int64, Int32, ... The default is Int.

chunk_size :: Int: Chunk size for HDF5. The default is 1000.

deflate :: Int: Deflate parameter for HDF5. The default is 3.

RationalPolygons.write_polygon_dataset — Function

write_polygon_dataset(f :: Union{HDF5.File, HDF5.Group}, path :: String, Ps :: Vector{RationalPolygon{T,N,M}}) where {N,M,T <: Integer}

Write the polygons Ps to an HDF5 dataset named path. Creates the dataset if it does not exist already. If it does exist, the data will be appended to it.

Example:

Write the reflexive lattice triangles to an HDF5 file:

julia> using RationalPolygons, StaticArrays, HDF5

julia> Vs = SMatrix{2,3}[[1 0 -2 ; 0 1 -1], [1 1 -3 ; 0 2 -4], [1 0 -1 ; 0 1 -1],[1 0 -3 ; 0 1 -2], [1 1 -2 ; 0 3 -3]];

julia> Ps = [RationalPolygon(V,1) for V ∈ Vs];

julia> f = h5open("/tmp/reflexive_triangles.h5", "cw")
🗂️ HDF5.File: (read-write) /tmp/reflexive_triangles.h5

julia> write_polygon_dataset(f, "my_triangles", Ps)

julia> close(f)

The resulting HDF5 file can be inspected using a command line tool such as h5dump:

$ h5dump /tmp/reflexive_triangles.h5 
HDF5 "/tmp/reflexive_triangles.h5" {
GROUP "/" {
   DATASET "my_triangles" {
      DATATYPE  H5T_COMPOUND {
         H5T_STD_I64LE "1";
         H5T_STD_I64LE "2";
         H5T_STD_I64LE "3";
         H5T_STD_I64LE "4";
         H5T_STD_I64LE "5";
         H5T_STD_I64LE "6";
      }
      DATASPACE  SIMPLE { ( 5 ) / ( H5S_UNLIMITED ) }
      DATA {
      (0): {
            1,
            0,
            0,
            1,
            -2,
            -1
         },
      (1): {
            1,
            0,
            1,
            2,
            -3,
            -4
         },
      (2): {
            1,
            0,
            0,
            1,
            -1,
            -1
         },
      (3): {
            1,
            0,
            0,
            1,
            -3,
            -2
         },
      (4): {
            1,
            0,
            1,
            3,
            -2,
            -3
         }
      }
   }
}
}

We can read them back into RationalPolygons.jl at any time using read_polygon_dataset:

julia> using RationalPolygons, HDF5

julia> f = h5open("/tmp/reflexive_triangles.h5", "r")
🗂️ HDF5.File: (read-only) /tmp/reflexive_triangles.h5
└─ 🔢 my_triangles

julia> Ps = read_polygon_dataset(1, f, "my_triangles")
5-element Vector{RationalPolygon{Int64, 3, 6}}:
 Rational polygon of rationality 1 with 3 vertices.
 Rational polygon of rationality 1 with 3 vertices.
 Rational polygon of rationality 1 with 3 vertices.
 Rational polygon of rationality 1 with 3 vertices.
 Rational polygon of rationality 1 with 3 vertices.

RationalPolygons.read_polygon_dataset — Function

read_polygon_dataset(k :: T, f :: Union{HDF5.File, HDF5.Group}, path :: String, I...) where {T <: Integer}

Read from an HDF5 dataset containing $k$-rational polygons.